Finalise refactor
