+++ thewalrus/tests/test_quantum.py
@@ -29,8 +29,8 @@ from thewalrus.quantum import (
     Xmat,
     Qmat,
     Amat,
-    Beta,
-    prefactor,
+    complex_to_real_displacements,
+    _prefactor,
     density_matrix_element,
     density_matrix,
     find_scaling_adjacency_matrix,
@@ -38,7 +38,7 @@ from thewalrus.quantum import (
     mean_number_of_clicks_graph,
     Covmat,
     gen_Qmat_from_graph,
-    Means,
+    real_to_complex_displacements,
     photon_number_covmat,
     is_valid_cov,
     is_pure_cov,
@@ -46,8 +46,8 @@ from thewalrus.quantum import (
     state_vector,
     is_classical_cov,
     find_classical_subsystem,
-    total_photon_num_dist_pure_state,
-    gen_single_mode_dist,
+    pure_state_distribution,
+    _squeezed_state_distribution,
     fock_tensor,
     photon_number_mean_vector,
     photon_number_mean,
@@ -180,7 +180,7 @@ def test_Amat_TMS_using_Q():
 def test_beta():
     """test the correct beta is returned"""
     mu = np.arange(4)
-    res = Beta(mu)
+    res = complex_to_real_displacements(mu)
 
     alpha = (mu[:2] + 1j * mu[2:]) / np.sqrt(2 * 2)
     ex = np.concatenate([alpha, alpha.conj()])
@@ -190,8 +190,8 @@ def test_beta():
 def test_Means():
     """test the correct beta is returned"""
     res = np.arange(4)
-    mu = Beta(res)
-    ex = Means(mu)
+    mu = complex_to_real_displacements(res)
+    ex = real_to_complex_displacements(mu)
     assert np.allclose(res, ex)
 
 
@@ -200,7 +200,7 @@ def test_prefactor_vacuum():
     Q = np.identity(2)
     beta = np.zeros([2])
 
-    res = prefactor(Means(beta), Covmat(Q))
+    res = _prefactor(real_to_complex_displacements(beta), Covmat(Q))
     ex = 1
     assert np.allclose(res, ex)
 
@@ -213,7 +213,7 @@ def test_prefactor_TMS():
 
     beta = np.zeros([4])
 
-    res = prefactor(Means(beta), Covmat(Q))
+    res = _prefactor(real_to_complex_displacements(beta), Covmat(Q))
     ex = 0.5
     assert np.allclose(res, ex)
 
@@ -228,7 +228,7 @@ def test_prefactor_with_displacement():
     vect = 1.2 * np.ones([2]) + 1j * np.ones(2)
     beta = np.concatenate([vect, vect.conj()])
 
-    res = prefactor(Means(beta), Covmat(Q), hbar=2)
+    res = _prefactor(real_to_complex_displacements(beta), Covmat(Q), hbar=2)
     ex = np.exp(-0.5 * beta @ Qinv @ beta.conj()) / np.sqrt(np.linalg.det(Q))
     assert np.allclose(res, ex)
 
@@ -240,18 +240,18 @@ def test_density_matrix_element_vacuum():
 
     el = [[0], [0]]
     ex = 1
-    res = density_matrix_element(Means(beta), Covmat(Q), el[0], el[1])
+    res = density_matrix_element(real_to_complex_displacements(beta), Covmat(Q), el[0], el[1])
     assert np.allclose(ex, res)
 
     el = [[1], [1]]
     #    res = density_matrix_element(beta, A, Q, el[0], el[1])
-    res = density_matrix_element(Means(beta), Covmat(Q), el[0], el[1])
+    res = density_matrix_element(real_to_complex_displacements(beta), Covmat(Q), el[0], el[1])
 
     assert np.allclose(0, res)
 
     el = [[1], [0]]
     #    res = density_matrix_element(beta, A, Q, el[0], el[1])
-    res = density_matrix_element(Means(beta), Covmat(Q), el[0], el[1])
+    res = density_matrix_element(real_to_complex_displacements(beta), Covmat(Q), el[0], el[1])
 
     assert np.allclose(0, res)
 
@@ -282,12 +282,12 @@ mu = np.array([0.04948628, -0.55738964, 0.71298259, 0.17728629, -0.14381673, 0.3
 @pytest.mark.parametrize("t", [t0, t1, t2, t3, t4])
 def test_density_matrix_element_disp(t):
     """Test density matrix elements for a state with displacement"""
-    beta = Beta(mu)
+    beta = complex_to_real_displacements(mu)
     Q = Qmat(V)
 
     el = t[0]
     ex = t[1]
-    res = density_matrix_element(Means(beta), Covmat(Q), el[0], el[1])
+    res = density_matrix_element(real_to_complex_displacements(beta), Covmat(Q), el[0], el[1])
     assert np.allclose(ex, res)
 
 
@@ -302,12 +302,12 @@ t4 = [[0, 2, 0], [0, 2, 3]], 0
 @pytest.mark.parametrize("t", [t0, t1, t2, t3, t4])
 def test_density_matrix_element_no_disp(t):
     """Test density matrix elements for a state with no displacement"""
-    beta = Beta(np.zeros([6]))
+    beta = complex_to_real_displacements(np.zeros([6]))
     Q = Qmat(V)
 
     el = t[0]
     ex = t[1]
-    res = density_matrix_element(Means(beta), Covmat(Q), el[0], el[1])
+    res = density_matrix_element(real_to_complex_displacements(beta), Covmat(Q), el[0], el[1])
     assert np.allclose(ex, res)
 
 
@@ -565,7 +565,7 @@ def test_pure_state_amplitude_coherent(i):
     """ Tests pure state amplitude for a coherent state """
     cov = np.identity(2)
     mu = np.array([1.0, 2.0])
-    beta = Beta(mu)
+    beta = complex_to_real_displacements(mu)
     alpha = beta[0]
     exact = np.exp(-0.5 * np.abs(alpha) ** 2) * alpha ** i / np.sqrt(np.math.factorial(i))
     num = pure_state_amplitude(mu, cov, [i])
@@ -628,7 +628,7 @@ def test_state_vector_coherent():
     cutoff = 5
     cov = np.identity(2)
     mu = np.array([1.0, 2.0])
-    beta = Beta(mu)
+    beta = complex_to_real_displacements(mu)
     alpha = beta[0]
     exact = np.array([(np.exp(-0.5 * np.abs(alpha) ** 2) * alpha ** i / np.sqrt(np.math.factorial(i))) for i in range(cutoff)])
     num = state_vector(mu, cov, cutoff=cutoff)
@@ -669,15 +669,15 @@ def test_is_classical_cov_thermal(nbar):
 
 
 @pytest.mark.parametrize("cutoff", [50, 51, 52, 53])
-def test_total_photon_num_dist_pure_state(cutoff):
+def test_pure_state_distribution(cutoff):
     """ Test the correct photon number distribution is obtained for n modes
     with nmean number of photons up to Fock cutoff nmax"""
     n = 3
     nmean = 1.0
     rs = np.arcsinh(np.sqrt(nmean)) * np.ones([n])
     cov = np.diag(np.concatenate([np.exp(2 * rs), np.exp(-2 * rs)]))
-    p1 = total_photon_num_dist_pure_state(cov, cutoff=cutoff)
-    p2 = gen_single_mode_dist(np.arcsinh(np.sqrt(nmean)), N=n, cutoff=cutoff)
+    p1 = pure_state_distribution(cov, cutoff=cutoff)
+    p2 = _squeezed_state_distribution(np.arcsinh(np.sqrt(nmean)), N=n, cutoff=cutoff)
     assert np.allclose(p1, p2)
 
 
@@ -1033,14 +1033,14 @@ def test_update_with_noise_coherent(num_modes, parallel, monkeypatch):
     noise_dists = np.array([poisson.pmf(np.arange(cutoff), nbar) for nbar in nbar_vals])
     hbar = 2
     beta = np.random.rand(num_modes) + 1j * np.random.rand(num_modes)
-    means = Means(np.concatenate((beta, beta.conj())), hbar=hbar)
+    means = real_to_complex_displacements(np.concatenate((beta, beta.conj())), hbar=hbar)
     cov = hbar * np.identity(2 * num_modes) / 2
     cutoff = 10
 
     probs = probabilities(means, cov, cutoff, parallel=parallel, hbar=2)
     updated_probs = update_probabilities_with_noise(noise_dists, probs)
     beta_expected = np.sqrt(nbar_vals + np.abs(beta) ** 2)
-    means_expected = Means(
+    means_expected = real_to_complex_displacements(
         np.concatenate((beta_expected, beta_expected.conj())), hbar=hbar
     )
     expected = probabilities(means_expected, cov, cutoff, parallel=parallel, hbar=2)
@@ -1055,7 +1055,7 @@ def test_update_with_noise_coherent_value_error():
     noise_dists = np.array([poisson.pmf(np.arange(cutoff), nbar) for nbar in nbar_vals])
     hbar = 2
     beta = np.random.rand(num_modes) + 1j * np.random.rand(num_modes)
-    means = Means(np.concatenate((beta, beta.conj())), hbar=hbar)
+    means = real_to_complex_displacements(np.concatenate((beta, beta.conj())), hbar=hbar)
     cov = hbar * np.identity(2 * num_modes) / 2
     cutoff = 10
     probs = probabilities(means, cov, cutoff, hbar=2)
@@ -1099,8 +1099,8 @@ def test_fidelity_coherent_state(num_modes, hbar):
     """Test the fidelity of two multimode coherent states"""
     beta1 = np.random.rand(num_modes) + 1j * np.random.rand(num_modes)
     beta2 = np.random.rand(num_modes) + 1j * np.random.rand(num_modes)
-    means1 = Means(np.concatenate([beta1, beta1.conj()]), hbar=hbar)
-    means2 = Means(np.concatenate([beta2, beta2.conj()]), hbar=hbar)
+    means1 = real_to_complex_displacements(np.concatenate([beta1, beta1.conj()]), hbar=hbar)
+    means2 = real_to_complex_displacements(np.concatenate([beta2, beta2.conj()]), hbar=hbar)
     cov1 = hbar * np.identity(2 * num_modes) / 2
     cov2 = hbar * np.identity(2 * num_modes) / 2
     fid = fidelity(means1, cov1, means2, cov2, hbar=hbar)
@@ -1114,7 +1114,7 @@ def test_fidelity_coherent_state(num_modes, hbar):
 def test_fidelity_vac_to_displaced_squeezed(r, alpha, hbar):
     """Calculates the fidelity between a coherent squeezed state and vacuum"""
     cov1 = np.diag([np.exp(2 * r), np.exp(-2 * r)]) * hbar / 2
-    means1 = Means(np.array([alpha, np.conj(alpha)]), hbar=hbar)
+    means1 = real_to_complex_displacements(np.array([alpha, np.conj(alpha)]), hbar=hbar)
     means2 = np.zeros([2])
     cov2 = np.identity(2) * hbar / 2
     expected = (
@@ -1152,7 +1152,7 @@ def test_fidelity_wrong_shape():
 def test_expectation_normal_ordered_coherent(alpha, hbar):
     """Test the correct evaluation of the normal ordered expectation value for a product of coherent states"""
     beta = np.concatenate([alpha, np.conj(alpha)])
-    means = Means(beta, hbar=hbar)
+    means = real_to_complex_displacements(beta, hbar=hbar)
     cov = np.identity(len(beta)) * hbar / 2
     pattern = np.random.randint(low=0, high=3, size=len(beta))
     result = normal_ordered_expectation(means, cov, pattern, hbar=hbar)
@@ -1219,7 +1219,7 @@ def test_single_mode_displaced_squeezed(r, phi, x, y):
     cov = 0.5 * hbar * S @ S.T
     beta = np.array([x + 1j * y, x - 1j * y])
     alpha = beta[0]
-    means = Means(beta, hbar=hbar)
+    means = real_to_complex_displacements(beta, hbar=hbar)
     a = alpha
     adxa = np.abs(alpha) ** 2 + np.sinh(r) ** 2
     a2 = alpha ** 2 - np.exp(1j * phi) * 0.5 * np.sinh(2 * r)
@@ -1267,7 +1267,7 @@ def test_expt_two_mode_squeezed(r, phi):
 def test_photon_number_expectation_displaced(alpha, hbar):
     """Tests the correct photon number expectation for coherent states"""
     beta = np.concatenate([alpha, np.conj(alpha)])
-    means = Means(beta, hbar=hbar)
+    means = real_to_complex_displacements(beta, hbar=hbar)
     cov = np.identity(len(beta)) * hbar / 2
     val = photon_number_expectation(
         means, cov, modes=list(range(len(alpha))), hbar=hbar
